requirements.txt
streamlit
openai
python-dotenv
pandas
pypdf
python-docx
